Refactor token-rule remove to not use user input
- The first step of /starry token-rule remove now provides a select
  dropdown instead of requiring the user to type in the exact name
  of the role they'd like to remove.
- The removeConfirmation and removeRejection steps have been
  consolidated into a single removeRole step. The removeVerify step
  will always navigate to removeRole now, as removeRole will check
  which button was actually pressed by the user.
