--- src/commands/tokenRemove/removeRejection.js
@@ -1,5 +1,0 @@
-module.exports = {
-  removeRejection: {
-    message: '✨ 👍 🌟',
-  }
-}
